requirements.txt
aiogram==3.2.0
aiohttp==3.9.3
aiosqlite==0.19.0
pydantic-settings==2.2.1
transformers==4.39.3
torch==2.2.2
Pillow==10.2.0
